Moved unit-test to test_movielens.py
